elearn/elearn/src/test/java/com/application/demo/controller/UserControllerTest.java
package com.application.demo.controller;

import static org.mockito.Mockito.*;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import com.application.demo.Dto.UseTempDto;
import com.application.demo.repository.UserFullDetailsRepository;
import com.application.demo.repository.UserTempRepository;
import com.application.demo.service.EmailService;
import com.application.demo.service.UserService;

@ExtendWith(SpringExtension.class)
@WebMvcTest(controllers = UserController.class)
@AutoConfigureMockMvc
public class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private EmailService emailService;

    @MockBean
    private UserService userService;
    
    @MockBean
    private UserTempRepository userTempRepository;
    
    @MockBean
    private UserFullDetailsRepository userFullDetailsRepository;

    // Test for createStudent API success scenario
    @Test
    public void testCreateStudent_Success() throws Exception {
        // Mock the necessary dependencies
        // For simplicity, you can use Mockito to mock the behavior of the repositories and services

        mockMvc.perform(MockMvcRequestBuilders.post("/api/admin/create-student")
                .contentType(MediaType.APPLICATION_JSON)
                .content("{ \"email\": \"talapaneni142@gmail.com\", \"role\": \"teacher\" }"))
                .andExpect(MockMvcResultMatchers.status().isCreated())
                .andExpect(MockMvcResultMatchers.content().string("Student registration initiated successfully."));
    }

    // Test for createStudent API failure scenario
    @Test
    public void testCreateStudent_Failure() throws Exception {
        // Mock the necessary dependencies to simulate an error during registration

        mockMvc.perform(MockMvcRequestBuilders.post("/api/admin/create-student")
                .contentType(MediaType.APPLICATION_JSON)
                .content("{ \"email\": \"talapaneni142@gmail.com\", \"role\": \"teacher\" }"))
                .andExpect(MockMvcResultMatchers.status().isInternalServerError())
                .andExpect(MockMvcResultMatchers.content().string("An error occurred."));
    }

    // Repeat the above structure for other APIs (completeRegistration, sentMailRequest, sendOtp, resetPassword) with success and failure scenarios.
}
